fypScraper/requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
cachetools==2.0.1
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
dill==0.2.7.1
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
google-api-core==0.1.4
google-auth==1.3.0
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
grpcio==1.9.0
httplib2==0.10.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
lxml==4.1.1
oauth2client==3.0.0
parsel==1.4.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
pytz==2017.3
queuelib==1.4.2
requests==2.18.4
rsa==3.4.2
Scrapy==1.6.0
scrapy-splash==0.7.2
selenium==3.11.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
urllib3==1.22
w3lib==1.19.0
zope.interface==4.4.3
